services/notifier/requirements.txt
sendgrid>=6.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
